dec 4 work
this was a fun one, i had complicated it by trying to dynamically append to my slice but its simpler to just count the lines and fill it out prior to processing. It makes more sense as well because I know I have 1 of at least each card at the start anyhow.